Fixed some datatype bugs, changed auth controller param type, removed un used code

+++ src/main/java/com/recommit/assignment/formula1/formula1dataservice/dto/ergastApiResponse/DriverStandingsDTO.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class DriverStandingsDTO implements Serializable {
     private Integer position;
     private String positionText;
-    private Integer points;
+    private Float points;
     private Integer wins;
     @JsonProperty("Driver")
     private DriverDTO driver;
